Replace ad-hoc stderr announcements with `log` API calls.
This makes things tidier by default, and reflects how mustang is
becoming more usable.

Also, reorganize the code to have a cleaner module structure, now
that the module structure is exposed in log messages.
