+++ mustang/README.md
@@ -1,5 +1,5 @@
 This is the main public interface to `mustang`. It automatically
-pulls in crates to perform process initialization and termination,
+pulls in crates to perform program initialization and termination,
 provide selected libc symbols, and registers a global allocator.
 
 To use, add a `mustang` dependency to Cargo.toml, and add this line
