+++ origin/Cargo.toml
@@ -15,6 +15,12 @@ linux-raw-sys = "0.0.28"
 rsix = "0.23.6"
 bitflags = "1.3.0"
 once_cell = "1.8.0"
+log = "0.4.14"
+
+# Optional logging backend. You can use any external logger, but using this
+# feature allows origin to initialize the logger before main, so that you can
+# see the log messages emitted before main is called.
+env_logger = { version = "0.9.0", optional = true }
 
 [features]
 default = ["threads"]
