+++ origin/src/lib.rs
@@ -6,8 +6,7 @@
 #![feature(atomic_mut_ptr)]
 #![cfg(target_vendor = "mustang")]
 
-mod entry;
-mod process;
+mod program;
 #[cfg(feature = "threads")]
 mod threads;
 
@@ -24,10 +23,9 @@ mod arch;
 #[path = "arch-riscv64.rs"]
 mod arch;
 
-use entry::program;
 use std::ffi::c_void;
 
-pub use process::{at_exit, exit, exit_immediately};
+pub use program::{at_exit, exit, exit_immediately};
 #[cfg(feature = "threads")]
 pub use threads::{
     at_thread_exit, create_thread, current_thread, current_thread_id, default_guard_size,
@@ -45,6 +43,8 @@ pub use threads::{
 #[link_section = ".text.__mustang"]
 #[no_mangle]
 unsafe extern "C" fn _start() -> ! {
+    use program::entry;
+
     // Jump to `program`, passing it the initial stack pointer value as an
     // argument, a NULL return address, a NULL frame pointer, and an aligned
     // stack pointer.
@@ -52,23 +52,23 @@ unsafe extern "C" fn _start() -> ! {
     #[cfg(target_arch = "x86_64")]
     asm!("mov rdi, rsp",
          "push rbp",
-         "jmp {program}",
-         program = sym program,
+         "jmp {entry}",
+         entry = sym entry,
          options(noreturn));
 
     #[cfg(target_arch = "aarch64")]
     asm!("mov x0, sp",
          "mov x30, xzr",
-         "b {program}",
-         program = sym program,
+         "b {entry}",
+         entry = sym entry,
          options(noreturn));
 
     #[cfg(target_arch = "riscv64")]
     asm!("mv a0, sp",
          "mv ra, zero",
          "mv fp, zero",
-         "tail {program}",
-         program = sym program,
+         "tail {entry}",
+         entry = sym entry,
          options(noreturn));
 
     #[cfg(target_arch = "x86")]
@@ -78,8 +78,8 @@ unsafe extern "C" fn _start() -> ! {
          "push ebp",
          "push eax",
          "push ebp",
-         "jmp {program}",
-         program = sym program,
+         "jmp {entry}",
+         entry = sym entry,
          options(noreturn));
 }
 
@@ -94,6 +94,25 @@ unsafe impl Sync for SendSyncVoidStar {}
 #[no_mangle]
 #[used]
 static __dso_handle: SendSyncVoidStar = SendSyncVoidStar(&__dso_handle as *const _ as *mut c_void);
+
+/// Initialize logging, if enabled.
+#[cfg(feature = "env_logger")]
+#[link_section = ".init_array.00099"]
+#[used]
+static INIT_ARRAY: unsafe extern "C" fn() = {
+    unsafe extern "C" fn function() {
+        env_logger::init();
+
+        log::trace!(target: "origin::program", "Program started");
+
+        // Log the thread id. We initialized the main earlier than this, but
+        // we couldn't initialize the logger until after the main thread is
+        // intialized :-).
+        #[cfg(feature = "threads")]
+        log::trace!(target: "origin::threads", "Main Thread[{:?}] initialized", current_thread_id());
+    }
+    function
+};
 
 /// Ensure that this module is linked in.
 #[inline(never)]
